Anon stem progress

+++ .anon.dir/Anon/aard.php
@@ -14,12 +14,13 @@ namespace Anon;
       {
           permit::fubu('clan:sudo,lead,gang,geek'); $post=knob($_POST);
           if(!$purl){$purl=$post->purl; $vars=$post->vars;};
-          expect::purl($purl); $info=path::info($purl); $addr="$info->plug://$info->host";
-dump($addr);
+          expect::purl($purl); $info=path::info($purl); $addr="https://$info->host";
+
           $code = impose(pget('$/Anon/base/deploy.php'),'{:',':}',$vars);
           $plug = plug($purl);  $plug->vivify();
 
-          if($plug->link){signal::busy(['with'=>'remoteDeploy','done'=>10]);};
+          if(!$plug->link){$f=$plug->fail; if(!$f){$f="connection failure";}; fail::remoteDeploy($f); exit;};
+          signal::busy(['with'=>'remoteDeploy','done'=>10]);
 
           $done = $plug->delete(['.htaccess','index.php','index.html']);
           if(!$done){fail::remoteDeploy('unable to delete remote auto-handler'); exit;};
@@ -27,6 +28,9 @@ dump($addr);
 
           $done = $plug->insert(['index.php'=>$code]);
           if(!$done){fail::remoteDeploy('unable to insert remote auto-handler'); exit;};
+          signal::busy(['with'=>'remoteDeploy','done'=>20]);
+
+          return $addr;
       }
    }
 # ---------------------------------------------------------------------------------------------------------------------------------------------
